Delete only posts belonging to logged in user
implemented function for asking the post for its owner and only if the owner is the current user, he is able to delete post.

override equals() and hashcode() methods for comparing objects with are the same by attributes but differ on the hash.

+++ src/main/java/uni/kassel/web17/spring/post/PostService.java
@@ -17,8 +17,6 @@ public class PostService {
 	private PostRepo postRepo;
     @Autowired
     private UserService userService;
-    @Autowired
-    private UserRepo userRepo;
 
 	public Iterable<PostObj> getPosts() {
 		User currentUser = userService.getCurrentUser();
@@ -42,6 +40,11 @@ public class PostService {
 	}
 
 	public void deletePostById(int id) {
+		// Validate that user is allowed to delete post.
+		PostObj post = postRepo.findOne(id);
+		if (!post.getAuthor().equals(userService.getCurrentUser())) {
+			throw new IllegalStateException("User not allowed to delete post");
+		}
 
 		postRepo.delete(id);
 	}
